pyproject.toml
[tool.poetry]
name = "test-technique"
version = "0.1.0"
description = ""
authors = ["Quicksign"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pillow = "^11.3.0"
transformers = "^4.53.1"
torch = "^2.7.1"
pydantic = "^2.11.7"
python-multipart = "^0.0.20"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
httpx = "*"
ruff= "*"
mypy = "*"

[tool.coverage.report]
show_missing = true
fail_under = 99
omit = ["**/tests/test_*.py","app/main.py"]

[tool.mypy]
python_version = "3.10"
namespace_packages = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 82

[tool.ruff.lint]
select = [
  "F",
  "E",
  "D",
  "I"
]
ignore = [
  "E501"
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"

[tool.ruff.lint.pydocstyle]
convention = "google"


[tools.poetry.dev]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
